feat(apps): share PNM/PGX/TIFF readers via openhtj2k_imgio + add streaming TIFF input
- New static lib `openhtj2k_imgio` (source/apps/imgio/) holds the PNM/PGX/TIFF
  stream readers plus an `imgio::open_stream_reader()` factory. Linked PRIVATE
  by the encoder and imgcmp; not part of the public libopenhtj2k ABI, and
  `OPENHTJ2K_TIFF_SUPPORT` is PRIVATE on the new target only.
- Encoder: streaming (default, line-based) path now accepts `.tif` / `.tiff`
  inputs (8/16-bit, RGB or grayscale, both planar configurations,
  uncompressed or strip-compressed). Tiled TIFFs still require `-batch`.
- imgcmp: migrated to imgio. Reformulated PAE/MSE/PSNR as line-based
  reductions over per-component planar rows (mathematically identical to the
  prior interleaved-buffer loop; output format byte-exact). Adds equality
  checks on num_components / bit_depth, gains TIFF input, drops
  `image_class.hpp`. Working-set drops from O(W·H·nc·4) to O(nc·W·4) bytes
  per side — enables gigapixel comparisons.
- PGX parser: fix latent trailing-whitespace overrun. The post-header
  "skip single trailing whitespace" was a `while`-loop that greedily
  consumed data bytes equal to 0x09/0x0a/0x0d/0x20, shifting every row by
  one byte and overrunning EOF on the last fread. Per spec, exactly one
  whitespace separates header from data — consume one, no more. Same class
  of bug as the documented PnmStreamReader fix.
- Core lib `image::read_pnmpgx`: replace `fgets(buf, 256, fp)` with
  `do { fgetc } while (… != \n …)` so PNM comments longer than 255 chars
  are no longer truncated (truncation caused the next byte to be parsed
  mid-comment as header data, producing garbage width/height/maxval).
- Docs: README format-support row and `docs/cli_encoder.md` opening
  paragraph updated to reflect that streaming TIFF input is now supported.

Tests: 627/627 native (`ctest --build-config Release -j 8`),
       137/137 WASM tests, gigapixel TIFF round-trip (encode + reduce-4
       decode) byte-identical to pre-refactor output.
