+++ source/apps/imgio/CMakeLists.txt
@@ -1,0 +1,5 @@
+cmake_policy(SET CMP0076 NEW)
+target_sources(openhtj2k_imgio
+    PRIVATE
+    imgio.cpp
+)
